[refactor] create review response model
