+++ app/src/main/java/com/mashup/zuzu/data/response/model/IsBitter.kt
@@ -1,0 +1,12 @@
+package com.mashup.zuzu.data.response.model
+
+import com.google.gson.annotations.SerializedName
+
+data class IsBitter(
+
+    @SerializedName("Bitter")
+    val Bitter: Int,
+
+    @SerializedName("Sweet")
+    val Sweet: Int
+)
