+++ app/src/main/java/com/mashup/zuzu/data/response/model/IsBurning.kt
@@ -1,0 +1,12 @@
+package com.mashup.zuzu.data.response.model
+
+import com.google.gson.annotations.SerializedName
+
+data class IsBurning(
+
+    @SerializedName("Burning")
+    val Burning: Int,
+
+    @SerializedName("Smooth")
+    val Smooth: Int
+)
