+++ app/src/main/java/com/mashup/zuzu/data/response/model/IsHeavy.kt
@@ -1,0 +1,12 @@
+package com.mashup.zuzu.data.response.model
+
+import com.google.gson.annotations.SerializedName
+
+data class IsHeavy(
+
+    @SerializedName("Heavy")
+    val Heavy: Int,
+
+    @SerializedName("Light")
+    val Light: Int
+)
