+++ app/src/main/java/com/mashup/zuzu/data/response/model/IsStrong.kt
@@ -1,0 +1,12 @@
+package com.mashup.zuzu.data.response.model
+
+import com.google.gson.annotations.SerializedName
+
+data class IsStrong(
+
+    @SerializedName("Mild")
+    val Mild: Int,
+
+    @SerializedName("Strong")
+    val Strong: Int
+)
