+++ app/src/main/java/com/mashup/zuzu/data/response/model/Taste.kt
@@ -1,0 +1,12 @@
+package com.mashup.zuzu.data.response.model
+
+import com.google.gson.annotations.SerializedName
+
+data class Taste(
+
+    @SerializedName("percent")
+    val percent: Int,
+
+    @SerializedName("tasteName")
+    val tasteName: String
+)
